// Bsuit/bsuite-backend/bsuite-backend/src/common/enum/account-type.enum.ts
export enum AccountType {
  ASSET = "Asset",
  EXPENSE = "Expense",
  INCOME = "Income",
  LIABILITY = "Liability",
}

export enum AccountTypeReport {
  ASSET = "Asset",
  EXPENSE = "Expense",
  INCOME = "Income",
  LIABILITY = "Liability",
  CONTACT = "Contact",
}

export enum AccountTypeTransact {
  ASSET = "Asset",
  EXPENSE = "Expense",
  INCOME = "Income",
  LIABILITY = "Liability",
  CONTACT = "Contact",
  TAX = "Tax",
}